service_default/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.8.0 ; python_version >= "3.12" and python_version < "4.0"
cachetools==5.5.2 ; python_version >= "3.12" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.12" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.12" and python_version < "4.0"
construct-typing==0.5.6 ; python_version >= "3.12" and python_version < "4.0"
construct==2.10.68 ; python_version >= "3.12" and python_version < "4.0"
deprecated==1.2.18 ; python_version >= "3.12" and python_version < "4.0"
fastapi-gae-logging==0.0.9 ; python_version >= "3.12" and python_version < "4.0"
fastapi==0.115.8 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.24.1 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.24.1 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.38.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-appengine-logging==1.6.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-audit-log==0.3.0 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-core==2.4.2 ; python_version >= "3.12" and python_version < "4.0"
google-cloud-logging==3.11.4 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.68.0 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos[grpc]==1.68.0 ; python_version >= "3.12" and python_version < "4.0"
grpc-google-iam-v1==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.70.0 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.70.0 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4.0"
httpcore==1.0.7 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.28.1 ; python_version >= "3.12" and python_version < "4.0"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
importlib-metadata==8.5.0 ; python_version >= "3.12" and python_version < "4.0"
jsonalias==0.1.1 ; python_version >= "3.12" and python_version < "4.0"
opentelemetry-api==1.30.0 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.26.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.29.3 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.27.2 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.10.6 ; python_version >= "3.12" and python_version < "4.0"
python-telegram-bot==21.10 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
retry-reloaded==0.0.8 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4.0"
solana==0.36.6 ; python_version >= "3.12" and python_version < "4.0"
solders==0.26.0 ; python_version >= "3.12" and python_version < "4.0"
starlette==0.45.3 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.12" and python_version < "4.0"
websockets==15.0 ; python_version >= "3.12" and python_version < "4.0"
wrapt==1.17.2 ; python_version >= "3.12" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.12" and python_version < "4.0"
